requirements.txt
omegaconf==2.3.0
Pillow==10.1.0
einops==0.7.0
git+https://github.com/tatsy/torchmcubes.git
transformers==4.35.0
trimesh==4.0.5
rembg
huggingface-hub
imageio[ffmpeg]
gradio
easydict
kornia
matplotlib
opencv-python
timm
wandb
xformers
pyYaml
lpips
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
pytorch-fid